API/python/requirements.txt
blinker
simplejson
watchdog
requests
Flask
google-cloud-datastore
requests
oauthlib
pyOpenSSL
Flask-Login
google-cloud-storage
pyyaml
GoogleAppEngineCloudStorageClient